requirements.txt
flask
Werkzeug
moviepy
flask-cors
requests
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
SpeechRecognition
google-cloud-speech
openai
python-dotenv
pydub
numpy
ffmpeg-python
emoji
opencv-python-headless
google-cloud-core
gunicorn
google-cloud-storage==2.9.0
google-auth
psutil
google-cloud-videointelligence
google-cloud-storage==2.9.0